# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Lighter Trading Bot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /config:
    get:
      operationId: getBotConfig
      tags: [config]
      summary: Get current bot configuration
      responses:
        "200":
          description: Bot configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
    put:
      operationId: updateBotConfig
      tags: [config]
      summary: Update bot configuration (API keys, account)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBotConfigRequest"
      responses:
        "200":
          description: Updated configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfig"
  /config/strategies:
    get:
      operationId: getStrategies
      tags: [config]
      summary: Get all configured strategies
      responses:
        "200":
          description: List of strategies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StrategiesResponse"
    post:
      operationId: createStrategy
      tags: [config]
      summary: Create a new strategy (DCA or Grid)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStrategyRequest"
      responses:
        "201":
          description: Created strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /config/strategies/{id}:
    put:
      operationId: updateStrategy
      tags: [config]
      summary: Update a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStrategyRequest"
      responses:
        "200":
          description: Updated strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
    delete:
      operationId: deleteStrategy
      tags: [config]
      summary: Delete a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
...
  /bot/stop/{strategyId}:
    post:
      operationId: stopBot
      tags: [bot]
      summary: Stop a strategy bot
      parameters:
        - name: strategyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Bot stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotStatusResponse"
...
components:
  schemas:
...
    BotConfig:
      type: object
      properties:
        accountIndex:
          type: integer
          nullable: true
        apiKeyIndex:
          type: integer
          nullable: true
        hasPrivateKey:
          type: boolean
        network:
          type: string
          enum: [mainnet, testnet]
        l1Address:
          type: string
          nullable: true
        notifyOnBuy:
          type: boolean
          nullable: true
        notifyOnSell:
          type: boolean
          nullable: true
        notifyOnError:
          type: boolean
          nullable: true
        notifyOnStart:
          type: boolean
          nullable: true
        notifyOnStop:
          type: boolean
          nullable: true
        hasNotifyBotToken:
          type: boolean
        notifyChatId:
          type: string
          nullable: true
      required: [hasPrivateKey, network]
    UpdateBotConfigRequest:
      type: object
      properties:
        accountIndex:
          type: integer
          nullable: true
        apiKeyIndex:
          type: integer
          nullable: true
        privateKey:
          type: string
          nullable: true
        network:
          type: string
          enum: [mainnet, testnet]
        l1Address:
          type: string
          nullable: true
        notifyBotToken:
          type: string
          nullable: true
        notifyChatId:
          type: string
          nullable: true
        notifyOnBuy:
          type: boolean
          nullable: true
        notifyOnSell:
          type: boolean
          nullable: true
        notifyOnError:
          type: boolean
          nullable: true
        notifyOnStart:
          type: boolean
          nullable: true
        notifyOnStop:
          type: boolean
          nullable: true
    StrategiesResponse:
      type: object
      properties:
        strategies:
          type: array
          items:
            $ref: "#/components/schemas/Strategy"
      required: [strategies]
    Strategy:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [dca, grid]
        marketIndex:
          type: integer
        marketSymbol:
          type: string
        isActive:
          type: boolean
        isRunning:
          type: boolean
        dcaConfig:
          $ref: "#/components/schemas/DcaConfig"
          nullable: true
        gridConfig:
          $ref: "#/components/schemas/GridConfig"
          nullable: true
        stats:
          $ref: "#/components/schemas/StrategyStats"
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, name, type, marketIndex, marketSymbol, isActive, isRunning, createdAt, updatedAt]
    DcaConfig:
      type: object
      properties:
        amountPerOrder:
          type: number
          description: USDC amount per DCA order
        intervalMinutes:
          type: integer
          description: Minutes between orders
        maxOrders:
          type: integer
          nullable: true
          description: Max number of orders (null = unlimited)
        side:
          type: string
          enum: [buy, sell]
        orderType:
          type: string
          enum: [market, limit]
        limitPriceOffset:
          type: number
          nullable: true
          description: Percentage offset from market price for limit orders
      required: [amountPerOrder, intervalMinutes, side, orderType]
    GridConfig:
      type: object
      properties:
        lowerPrice:
          type: number
        upperPrice:
          type: number
        gridLevels:
          type: integer
          description: Number of grid levels
        amountPerGrid:
          type: number
          description: USDC amount per grid level
        mode:
          type: string
          enum: [neutral, long, short]
        orderType:
          type: string
          enum: [market, limit]
        limitPriceOffset:
          type: number
          nullable: true
          description: USDC offset from crossing price for limit orders
        stopLoss:
          type: number
          nullable: true
          description: Stop loss price - bot stops if price falls below this
        takeProfit:
          type: number
          nullable: true
          description: Take profit price - bot stops if price rises above this
      required: [lowerPrice, upperPrice, gridLevels, amountPerGrid, mode, orderType]
    StrategyStats:
      type: object
      properties:
        totalOrders:
          type: integer
        successfulOrders:
          type: integer
        totalBought:
          type: number
        totalSold:
          type: number
        avgBuyPrice:
          type: number
        avgSellPrice:
          type: number
        realizedPnl:
          type: number
        unrealizedPnl:
          type: number
      required: [totalOrders, successfulOrders, totalBought, totalSold, realizedPnl, unrealizedPnl]
    CreateStrategyRequest:
      type: object
      properties:
        name:
          type: string
        type:
          type: string
          enum: [dca, grid]
        marketIndex:
          type: integer
        dcaConfig:
          $ref: "#/components/schemas/DcaConfig"
          nullable: true
        gridConfig:
          $ref: "#/components/schemas/GridConfig"
          nullable: true
      required: [name, type, marketIndex]
    UpdateStrategyRequest:
      type: object
      properties:
        name:
          type: string
        isActive:
          type: boolean
        dcaConfig:
          $ref: "#/components/schemas/DcaConfig"
          nullable: true
        gridConfig:
          $ref: "#/components/schemas/GridConfig"
          nullable: true
    DeleteResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]
    BotStatusResponse:
      type: object
      properties:
        strategyId:
          type: integer
        isRunning:
          type: boolean
        message:
          type: string
        nextRunAt:
          type: string
          nullable: true
      required: [strategyId, isRunning, message]
...